pyproject.toml
[tool.poetry]
name = "automaticinsights"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.235"
pydantic-yaml = "^1.0.0"
openai = "^0.27.8"
pydantic-computed = "^0.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
ipython = "^8.14.0"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
